Organizando email service e atualizando git ignore
