+++ backend-rupay/.gitignore
@@ -1,6 +1,5 @@
 HELP.md
 target/
-resources/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
