--- backend-rupay/src/main/java/com/pc1/backendrupay/email/EmailService.java
@@ -1,8 +1,0 @@
-package com.pc1.backendrupay.email;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface EmailService {
-    void sendSimpleMessage(String to, String subject, String text);
-}
