+++ backend-rupay/src/main/resources/application.properties
@@ -1,0 +1,22 @@
+server.port = 8081
+
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://localhost:5433/backendrupay
+spring.datasource.username=postgres
+spring.datasource.password=admin
+
+spring.jpa.properties..hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+application.security.jwt.secret-key=7B83EEBB84AB676EAC3BED71613708301E65A7E834D96CA57A8F6FEF3AB1551E
+application.security.jwt.expiration=86400000
+application.security.jwt.refresh-token.expiration=604800000
+
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.username=rupay.cg@gmail.com
+spring.mail.password=ndrdawjisvmzjgqe 
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
